+++ controllers/blobs_router.py
@@ -28,10 +28,12 @@ class BlobsByActivityRequestDto(BaseModel):
 
 @router.get("/all", response_model=list[BlobStatsDto])
 def get_all(
-    name_search: str | None = None, show_dead: bool = False
+    name_search: str | None = None, show_dead: bool = False, league_id: int | None = None
 ) -> list[BlobStatsDto]:
     try:
-        return get_all_blobs(name_search=name_search, show_dead=show_dead)
+        return get_all_blobs(
+            name_search=name_search, show_dead=show_dead, league_id=league_id
+        )
     except Exception as e:
         traceback.print_exc()
         raise HTTPException(status_code=500, detail=f"{e.with_traceback(None)}")